# apps/backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: WH Logique API
  version: 0.2.0
paths:
  /api/v1/locations:
    get:
      summary: List locations
      description: Returns all seeded warehouse locations without pagination.
      tags: [locations]
      responses:
        "200":
          description: Locations retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/items:
    post:
      summary: Create item
      description: Creates an item after trimming fields and normalizing its SKU to uppercase.
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemRequest"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          description: SKU already exists, including when the original item was soft-deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      summary: List items
      description: Returns non-deleted items using optional category filtering and pagination.
      tags: [items]
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ItemCategory"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Items retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/items/{id}:
    parameters:
      - $ref: "#/components/parameters/ItemID"
    get:
      summary: Get item
      description: Returns a non-deleted item by ID.
      tags: [items]
      responses:
        "200":
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/ItemNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      summary: Update item
      description: Fully replaces the SKU, name, category, and unit of a non-deleted item.
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemRequest"
      responses:
        "200":
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/ItemNotFound"
        "409":
          description: SKU already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete item
      description: Soft-deletes an active item while keeping its SKU reserved.
      tags: [items]
      responses:
        "200":
          description: Item deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteItemResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/ItemNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stock/receive:
    post:
      summary: Receive stock
      description: Applies one or more stock receipts atomically. Each line adds the given quantity to an item-location balance and records an append-only stock log entry.
      tags: [stock]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockReceiveRequest"
      responses:
        "200":
          description: Stock received successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          description: Item or location not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stock/{item_id}:
    parameters:
      - $ref: "#/components/parameters/StockItemID"
    get:
      summary: Get stock by item
      description: Returns every stock balance for a single item, grouped by location.
      tags: [stock]
      responses:
        "200":
          description: Stock retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/ItemNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stock/{item_id}/logs:
    parameters:
      - $ref: "#/components/parameters/StockItemID"
    get:
      summary: Get stock logs by item
      description: Returns every stock receipt for a single item, newest first.
      tags: [stock]
      responses:
        "200":
          description: Stock logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockLogListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    ItemID:
      name: id
      in: path
      required: true
      description: Item UUID
      schema:
        type: string
        format: uuid
    StockItemID:
      name: item_id
      in: path
      required: true
      description: Item UUID
      schema:
        type: string
        format: uuid
  responses:
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ItemNotFound:
      description: Item not found or has been soft-deleted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Location:
      type: object
      required: [id, code, zone, type]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: A-RACK
        zone:
          type: string
          enum: [A, B, C]
        type:
          type: string
          enum: [rack, shelf, bin, floor, cold_storage]
    LocationListResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Locations retrieved successfully
        data:
          type: array
          items:
            $ref: "#/components/schemas/Location"
    ItemCategory:
      type: string
      enum:
        - raw_material
        - finished_goods
        - packaging
        - spare_part
        - consumable
        - equipment
    ItemUnit:
      type: string
      enum: [pcs, kg, gr, ltr, box, carton, pallet]
    ItemRequest:
      type: object
      additionalProperties: false
      required: [sku, name, category, unit]
      properties:
        sku:
          type: string
          maxLength: 64
          example: rm-001
          description: Trimmed and normalized to uppercase before storage.
        name:
          type: string
          maxLength: 255
          example: Steel sheet
        category:
          $ref: "#/components/schemas/ItemCategory"
        unit:
          $ref: "#/components/schemas/ItemUnit"
    Item:
      type: object
      required: [id, sku, name, category, unit, created_at, deleted_at]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          example: RM-001
        name:
          type: string
          example: Steel sheet
        category:
          $ref: "#/components/schemas/ItemCategory"
        unit:
          $ref: "#/components/schemas/ItemUnit"
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    PaginationMeta:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 42
    ItemResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Item retrieved successfully
        data:
          $ref: "#/components/schemas/Item"
    ItemListResponse:
      type: object
      required: [success, message, data, meta]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Items retrieved successfully
        data:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    DeleteItemResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Item deleted successfully
    StockReceiveLine:
      type: object
      additionalProperties: false
      required: [item_id, location_id, qty]
      properties:
        item_id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
        qty:
          type: integer
          minimum: 1
    StockReceiveRequest:
      type: object
      additionalProperties: false
      required: [lines]
      properties:
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/StockReceiveLine"
    Stock:
      type: object
      required: [id, item_id, location_id, qty, updated_at]
      properties:
        id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
        qty:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time
    StockListResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Stock retrieved successfully
        data:
          type: array
          items:
            $ref: "#/components/schemas/Stock"
    StockLog:
      type: object
      required: [id, item_id, location_id, qty, created_at]
      properties:
        id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
        qty:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    StockLogListResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Stock logs retrieved successfully
        data:
          type: array
          items:
            $ref: "#/components/schemas/StockLog"
    ErrorDetail:
      type: object
      required: [reason]
      properties:
        field:
          type: string
        reason:
          type: string
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid request
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"